spark/docker-requirements.txt
google-cloud-dataproc==5.4.1
prefect-gcp[cloud_storage]==0.2.4
pyarrow==10.0.1
